+++ homehelper/changePassword_test.html
@@ -1,0 +1,57 @@
+<!DOCTYPE html>
+<!--
+Author: Code Apes.
+-->
+<html>
+<head>
+	<title>Account Information</title>
+	<meta charset="UTF-8">
+	<meta name="viewport" content="width=device-width, initial-scale=1.0">  
+	<link type="text/css" rel="stylesheet" href="styles/styles.css">
+	<link type="text/css" rel="stylesheet" href="styles/account.css">
+	<script>
+	</script>
+</head>
+<body>
+<?php
+require('header.php');
+?>
+<div class="container">
+	<!-- main body -->
+	<main>
+		<?php
+		require('account_sidebar.php');
+		?>
+		<div class="content" style="text-align:center;">
+		<h1>Change Password</h1>
+	<div id=">
+		<div style="background-color:transparent;">
+		<form method="POST" action="login.php">
+			<ul class="reg-box">
+				<li>
+					<label for="">Old Password</label>
+					<input type="password" name="oldpassword" value="" minlength="8" maxlength="16" style="color:#545454;"/>
+					<span class="error error1"></span>
+				</li>
+				<li>
+					<label for="">New Password</label>
+					<input type="password" name="newpassword" minlength="8" maxlength="16" style="color:#999;"/>
+					<span class="error error2"></span>
+				</li>
+			</ul>
+			
+			<div class="sub">
+				<input type="submit" value="Submit"/>
+			</div>
+		</form>
+		</div>
+	</div>
+		<br>
+		</div>
+	</main>
+</div>
+<?php
+require('footer.php');
+?>
+</body>
+</html>
