+++ homehelper/styles/account.css
@@ -1,73 +1,4 @@
 
-.container{padding:150px 120px 100px 100px;}
-
-.content {
-	margin-left:29%;
-	background-color: #FFFFFF;
-	color: #151517;
-	line-height: 200%;
-	overflow: auto; 
-}
-
-
-.sidebar a:hover{
-	background-color:#fff;
-	color:#EEC50F;
-}
-
-/* left float sidebar */
-.sidebar{
-	background-color:#EEC50F;
-	height:auto;
-	width:190px;
-	position:fixed!important;
-	z-index:1;
-	overflow:auto; 
-	font-weight: bold;
-    letter-spacing: 1px;
-	box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
-}
-
-.sidebar a{
-	background-color:#fff;
-	display: block; 
-	text-decoration: none; 
-	color: #EEC50F;
-	padding: 15px 20px;
-	border-top: 1px solid #F0F0F0;
-	font-size:15px;
-	text-align: center;
-}
-
-.sidebar a:hover{
-	box-shadow: 2px 2px 15px 0 rgba(0, 0, 0, 0.2);
-	color: #fff;
-	background-color: #EEC50F;
-}
-
-.sidebar h3{
-	color: #fff;
-	text-align: center;
-	padding:5px;
-}
-.content h1{
-	color: #EEC50F;
-}
-
-table {
-    border-collapse: collapse;
-    width: 100%;
-    border-bottom: 1px solid #ddd;
-}
-
-th, td {
-    padding: 8px;
-    text-align: left;
-    border-bottom: 1px solid #ddd;
-}
-
-tr:hover {background-color:#f5f5f5;}
-
 .btn, .btn-inverse{
 	display:inline-block; 
 	padding:3px 20px;
