+++ homehelper/userInfo_test.html
@@ -7,6 +7,7 @@ Author: Code Apes.
 	<title>Account Information</title>
 	<meta charset="UTF-8">
 	<meta name="viewport" content="width=device-width, initial-scale=1.0"> 
+	<link type="text/css" rel="stylesheet" href="styles/styles.css">
 	<link type="text/css" rel="stylesheet" href="styles/account.css">
 	<script>
 	function EditPhone(){
@@ -344,7 +345,7 @@ require('header.php');
 		require('account_sidebar.php');
 		?>
 		<div class="content" style="text-align:center;">
-		<h1>Personal Information</h1><hr>
+		<h1>Personal Information</h1>
 		<table>
 		<tr>
 		  <td><h3>Username:</h3></td>
